feat(gwas-ui-demo): refactor workflow list & add prefixed id
